refactor based on review comments
